pyproject.toml
[project]
name = "pulsetrace"
version = "0.1.0"
description = "XAI application"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "keras>=3.9.0",
    "lime>=0.2.0.1",
    "matplotlib>=3.10.1",
    "pandas>=2.2.3",
    "pydantic>=2.0",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.6.1",
    "shap>=0.47.0",
    "tensorflow>=2.18.0",
    "torch>=2.6.0",
]

[project.scripts]
pulsetrace = "pulsetrace.main:cli"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
